# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "All apis here",
    "version": "1.0.0",
    "title": "Reservation apis",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": ""
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "https://api.creat.sg",
  "basePath": "/api",
...
  "schemes": ["http"],
  "paths": {
...
    "/tablesection/seatingtype/{id}": {
      "post": {
        "tags": ["TableSection"],
        "description": "Create Table Section",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "seatingtype Id",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "color": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "minPax": {
                  "type": "number"
                },
                "maxPax": {
                  "type": "number"
                },
                "outletTable": {
                  "type": "array",
                  "items": {
                    "type": "number"
                  }
                }
              },
              "required": ["name", "color", "outletTable"]
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Create Table Section"
          }
        }
      },
      "get": {
        "tags": ["TableSection"],
        "description": "Get All Table Section",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "seatingtype Id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Get All Table Section"
          }
        }
      }
    },
    "/tablesection/{id}": {
      "delete": {
        "tags": ["TableSection"],
        "description": "Delete Table Section",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "tablesection Id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Delete Table Section"
          }
        }
      }
    },
    "/tablesection/private/seatingtype/{id}": {
      "post": {
        "tags": ["TableSection"],
        "description": "Create Private Section",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "seatingtype Id",
            "required": true
          },
          {
            "in": "formData",
            "name": "name",
            "type": "string",
            "required": true,
            "description": "name."
          },
          {
            "in": "formData",
            "name": "color",
            "type": "string",
            "required": true,
            "description": "color."
          },
          {
            "in": "formData",
            "name": "originalPrice",
            "type": "string",
            "required": true,
            "description": "originalPrice."
          },
          {
            "in": "formData",
            "name": "depositPrice",
            "type": "string",
            "required": true,
            "description": "depositPrice."
          },
          {
            "in": "formData",
            "name": "description",
            "type": "string",
            "required": true,
            "description": "description."
          },
          {
            "in": "formData",
            "name": "minPax",
            "type": "string",
            "required": true,
            "description": "minPax."
          },
          {
            "in": "formData",
            "name": "maxPax",
            "type": "string",
            "required": true,
            "description": "maxPax."
          },
          {
            "in": "formData",
            "name": "image",
            "type": "file",
            "required": true,
            "description": "image."
          },
          {
            "in": "formData",
            "name": "outletTables",
            "type": "string",
            "required": true,
            "description": "outletTables."
          }
        ],
        "responses": {
          "200": {
            "description": "Create Private Section"
          }
        }
      }
    },
    "/tablesection/private/{id}/seatingtype/{seatingtypeId}": {
      "put": {
        "tags": ["TableSection"],
        "description": "Update Private Section",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "Private Id",
            "required": true
          },
          {
            "name": "seatingtypeId",
            "in": "path",
            "type": "string",
            "description": "Seatingtype Id",
            "required": true
          },
          {
            "in": "formData",
            "name": "name",
            "type": "string",
            "required": true,
            "description": "name."
          },
          {
            "in": "formData",
            "name": "color",
            "type": "string",
            "required": true,
            "description": "color."
          },
          {
            "in": "formData",
            "name": "originalPrice",
            "type": "string",
            "required": true,
            "description": "originalPrice."
          },
          {
            "in": "formData",
            "name": "depositPrice",
            "type": "string",
            "required": true,
            "description": "depositPrice."
          },
          {
            "in": "formData",
            "name": "minPax",
            "type": "string",
            "required": true,
            "description": "minPax."
          },
          {
            "in": "formData",
            "name": "maxPax",
            "type": "string",
            "required": true,
            "description": "maxPax."
          },
          {
            "in": "formData",
            "name": "description",
            "type": "string",
            "required": true,
            "description": "description."
          },
          {
            "in": "formData",
            "name": "image",
            "type": "file",
            "required": true,
            "description": "image."
          },
          {
            "in": "formData",
            "name": "isActive",
            "type": "boolean",
            "required": true,
            "description": "isActive."
          }
        ],
        "responses": {
          "200": {
            "description": "Update Private Section"
          }
        }
      }
    },
    "/tablesection/private/outlet/{id}": {
      "get": {
        "tags": ["TableSection"],
        "description": "Get Private Table Section",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "outlet Id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Get Private Table Section"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization",
      "description": "Enter the token with the `Bearer: ` prefix, e.g. Bearer abcde12345  "
    }
  },
  "security": [{ "Bearer": [] }]
}